kmalloc_test: eliminate module and add to kmalloc
To avoid exposing internals of kmalloc for tests requiring that
internal knowledge.
