+++ src/kmalloc.c
@@ -7,18 +7,21 @@ See the file LICENSE for details.
 #include "kmalloc.h"
 #include "console.h"
 #include "kerneltypes.h"
-
-#define TESTING
-
-#ifndef TESTING
-#define STATIC static
-#else
-#define STATIC
-#endif
+#include "memorylayout.h"
 
 #define KUNIT sizeof(struct kmalloc_chunk)
 
-STATIC struct kmalloc_chunk *head = 0;
+#define KMALLOC_STATE_FREE 0xa1a1a1a1
+#define KMALLOC_STATE_USED 0xbfbfbfbf
+
+struct kmalloc_chunk {
+	int state;
+	int length;
+	struct kmalloc_chunk *next;
+	struct kmalloc_chunk *prev;
+};
+
+static struct kmalloc_chunk *head = 0;
 
 /*
 Initialize the linked list by creating a single chunk at
@@ -40,7 +43,7 @@ Split a large chunk into two, such that the current chunk
 has the desired length, and the next chunk has the remainder.
 */
 
-STATIC void ksplit( struct kmalloc_chunk *c, int length )
+static void ksplit( struct kmalloc_chunk *c, int length )
 {
 	struct kmalloc_chunk *n = (struct kmalloc_chunk *)((char *)c + length);
 
@@ -86,7 +89,7 @@ void * kmalloc( int length )
 	if(length-c->length > 2*KUNIT) {
 	       	ksplit(c,length);
 	}
-	
+
 	c->state = KMALLOC_STATE_USED;
 
 	// return a pointer to the memory following the chunk header
@@ -98,7 +101,7 @@ Attempt to merge a chunk with its successor,
 if it exists and both are in the free state.
 */
 
-STATIC void kmerge( struct kmalloc_chunk *c )
+static void kmerge( struct kmalloc_chunk *c )
 {
 	if(!c) return;
 
@@ -152,4 +155,69 @@ void kmalloc_debug()
 		}
 		console_printf("     %x %x %x %d\n",c,c->prev,c->next,c->length);
 	}
+}
+
+// Testing
+
+static void setup(void)
+{
+	kmalloc_init((char*)KMALLOC_START, KMALLOC_LENGTH);
+}
+
+static void tear_down(void)
+{
+}
+
+static int kmalloc_test_single_alloc(void)
+{
+	char *ptr = kmalloc(128);
+	struct kmalloc_chunk *next = 0;
+	int res = (unsigned long) ptr == (unsigned long) head + sizeof(struct kmalloc_chunk);
+	res &= head->state == KMALLOC_STATE_USED;
+	res &= head->length == 128 + sizeof(struct kmalloc_chunk);
+	res &= (char *) head->next == (char*) KMALLOC_START + head->length;
+	next = head->next;
+	res &= next->state == KMALLOC_STATE_FREE;
+	res &= next->length == KMALLOC_LENGTH - head->length;
+
+	return res;
+}
+
+static int kmalloc_test_single_alloc_and_free(void)
+{
+	char *ptr = kmalloc(128);
+	int res;
+	kfree(ptr);
+	res = head->state == KMALLOC_STATE_FREE;
+	res &= head->next == 0;
+	res &= head->length == KMALLOC_LENGTH;
+
+	return res;
+}
+
+int kmalloc_test(void)
+{
+	int (*tests[])(void) = {
+		kmalloc_test_single_alloc,
+		kmalloc_test_single_alloc_and_free,
+	};
+
+	int i = 0;
+	for (i = 0; i < sizeof(tests)/sizeof(tests[0]); i++)
+	{
+		console_printf("running test %d...", i);
+		int res;
+
+		setup();
+		res = tests[i]();
+		tear_down();
+
+		if (!res) {
+			console_printf("failed\n");
+			console_printf("\ntest %d failed.\n");
+			return 0;
+		}
+		console_printf("succeeded\n");
+	}
+	return 1;
 }
