+++ src/kmalloc.h
@@ -7,20 +7,11 @@ See the file LICENSE for details.
 #ifndef KMALLOC_H
 #define KMALLOC_H
 
-#define KMALLOC_STATE_FREE 0xa1a1a1a1
-#define KMALLOC_STATE_USED 0xbfbfbfbf
-
 void * kmalloc( int length );
 void   kfree( void *ptr );
 
 void kmalloc_init( char *start, int length );
 void kmalloc_debug();
-
-struct kmalloc_chunk {
-	int state;
-	int length;
-	struct kmalloc_chunk *next;
-	struct kmalloc_chunk *prev;
-};
+int kmalloc_test();
 
 #endif
