--- src/kmalloc_test.h
@@ -1,5 +1,0 @@
-#ifndef KMALLOC_TEST_H
-#define KMALLOC_TEST_H
-int kmalloc_test(void);
-
-#endif
